refactor: Rename `electron-vendors.config.json` to `.electron-vendors.json`
Just for normalizing files names
