+++ .github/workflows/update-electron-vendors.yml
@@ -41,4 +41,4 @@ jobs:
           commit-message: Update electron vendors
           branch: autoupdates/electron-vendors
           title: Update electron vendors
-          body: Updated versions of electron vendors in `electron-vendors.config.json` and `.browserslistrc` files
+          body: Updated versions of electron vendors in `.electron-vendors.cache.json` and `.browserslistrc` files
