+++ packages/renderer/vite.config.js
@@ -1,6 +1,6 @@
 /* eslint-env node */
 
-import {chrome} from '../../electron-vendors.config.json';
+import {chrome} from '../../.electron-vendors.cache.json';
 import {join} from 'path';
 import {builtinModules} from 'module';
 import vue from '@vitejs/plugin-vue';
